+++ src/app.js
@@ -3,52 +3,41 @@ require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const cookieParser = require('cookie-parser');
-const cors = require('cors');
+const helmet = require('helmet');
 const { errors } = require('celebrate');
-const movies = require('./routes/movies');
-const users = require('./routes/users');
+
 const index = require('./routes/index');
+
+const cors = require('./middlewares/cors');
+const limiter = require('./middlewares/rateLimiter');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const errorHandler = require('./middlewares/error');
-const authHandler = require('./middlewares/auth');
-const notFoundHandler = require('./middlewares/notfound');
+const notFoundHandler = require('./middlewares/notFound');
+
+const { devDbConnectionString } = require('./utils/constants');
 
 const { NODE_ENV, DB_CONNECTION_STRING, PORT = 3001 } = process.env;
 
 const app = express();
+app.set('trust proxy', 1);
 
-const defaultDbConnectionString = 'mongodb://localhost:27017/moviesdb';
-mongoose.connect(NODE_ENV === 'production' ? DB_CONNECTION_STRING || defaultDbConnectionString : defaultDbConnectionString, {
+mongoose.connect(NODE_ENV === 'production' ? DB_CONNECTION_STRING || devDbConnectionString : devDbConnectionString, {
   useNewUrlParser: true,
   useCreateIndex: true,
   useFindAndModify: false,
   useUnifiedTopology: true,
 });
 
-const options = {
-  origin: [
-    'http://localhost:3000',
-    'http://movies-explorer.jumpingcrab.com',
-    'https://movies-explorer.jumpingcrab.com',
-  ],
-  methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-  preflightContinue: false,
-  optionsSuccessStatus: 204,
-  credentials: true,
-};
-app.use('*', cors(options));
+app.use('*', cors);
+app.use(limiter);
 
+app.use(helmet());
 app.use(cookieParser());
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
 app.use(requestLogger);
 
 app.use('/', index);
-
-app.use(authHandler);
-
-app.use('/movies', movies);
-app.use('/users', users);
 
 app.use(errorLogger);
 app.use(errors());
