+++ src/utils/constants.js
@@ -1,1 +1,2 @@
 module.exports.devSecret = 'super-secret-key';
+module.exports.devDbConnectionString = 'mongodb://localhost:27017/moviesdb';
